Move ItemSlot and ItemBinding to shared package as it's not specific to characters
