+++ src/main/kotlin/com/kryszak/gwatlin/api/characters/model/character/equipment/EquipmentItem.kt
@@ -2,11 +2,13 @@ package com.kryszak.gwatlin.api.characters.model.character.equipment
 
 import com.google.gson.annotations.SerializedName
 import com.kryszak.gwatlin.api.items.model.stats.ItemStats
+import com.kryszak.gwatlin.api.shared.ItemSlot
+import com.kryszak.gwatlin.api.shared.ItemBinding
 
 data class EquipmentItem(
     val id: Int,
-    val slot: EquipmentItemSlot,
-    val binding: EquipmentItemBinding,
+    val slot: ItemSlot,
+    val binding: ItemBinding,
     @SerializedName("bound_to")
     val boundTo: String?,
     val infusions: List<Int>?,
